requirement.txt
# Version Specific Requirements for the project
openai==1.37.2
streamlit==1.35.0
python-dotenv==1.0.1

